Updated JMH plugin

+++ src/main/java/eu/plumbr/jpoint/Trade.java
@@ -3,10 +3,10 @@ package eu.plumbr.jpoint;
 public class Trade {
   public int ticket;
   public int amount;
-  public double price;
+  public int price;
   public boolean buy;
 
-  public Trade(int ticket, int amount, double price, boolean buy) {
+  public Trade(int ticket, int amount, int price, boolean buy) {
     this.ticket = ticket;
     this.amount = amount;
     this.price = price;
